restructure js
