+++ wanikani/static/js/session/constants.js
@@ -1,0 +1,8 @@
+export const TEST_OPTIONS = {
+  definition: 'definition',
+  pinyin: 'pinyin',
+}
+export const CHARACTER_FIELDS = {
+  definitions: 'definitions',
+  pinyin: 'pinyin',
+}
