+++ wanikani/templates/wanikani/base.html
@@ -10,7 +10,6 @@
       {% endblock %}
       {% block head_js %}
         <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/1.6.2/jquery.min.js"> </script>
-        <script type="text/javascript" src="{% static "js/script.js" %}"></script>
       {% endblock %}
     </head>
